backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.28.0
django-cors-headers==4.6.0
django-filter==24.3
django-celery-beat==2.7.0
psycopg[binary]==3.2.4
redis==5.2.1
celery==5.4.0
httpx==0.28.1
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.8.2
sentry-sdk==2.19.2
